requirements.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
et-xmlfile==1.1.0
Flask==2.3.2
Flask-Cors==3.0.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.24.3
opencv-python==4.7.0.72
openpyxl==3.1.2
pandas==2.0.2
Pillow==9.5.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tzdata==2023.3
waitress==2.1.2
Werkzeug==2.3.6